+++ doxyboiler.py
@@ -8,6 +8,8 @@ import shutil
 import CppHeaderParser 
 import jinja2 as j2
 
+from doxyboiler_utils import FileToLines, LinesToFile
+
 #cpph.classes['elem_t']['methods']['public'][0]
 #{'line_number': 18, 'unresolved_parameters': True, 'parent': {'inherits': [], 'line_number': 13, 'forward_declares': {'protected': [], 'public': [], 'private': []}, 'name': 'elem_t', 'parent': None, 'abstract': False, 'namespace': '', 'declaration_method': 'struct', 'properties': {'protected': [], 'public': [{'line_number': 14, 'constant': 0, 'reference': 0, 'raw_type': 'int', 'static': 0, 'array': 0, 'pointer': 0, 'aliases': [], 'typedef': None, 'namespace': '', 'function_pointer': 0, 'mutable': False, 'type': 'int', 'property_of_class': 'elem_t', 'parent': None, 'ctypes_type': 'ctypes.c_int', 'typedefs': 0, 'extern': False, 'class': 0, 'unresolved': False, 'name': 'int_a', 'fundamental': True}, {'line_number': 15, 'constant': 0, 'reference': 0, 'raw_type': 'float', 'static': 0, 'array': 0, 'pointer': 0, 'aliases': [], 'typedef': None, 'namespace': '', 'function_pointer': 0, 'mutable': False, 'type': 'float', 'property_of_class': 'elem_t', 'parent': None, 'ctypes_type': 'ctypes.c_float', 'typedefs': 0, 'extern': False, 'class': 0, 'unresolved': False, 'name': 'float_a', 'fundamental': True}, {'line_number': 16, 'constant': 0, 'reference': 0, 'raw_type': 'double', 'static': 0, 'array': 0, 'pointer': 0, 'aliases': [], 'typedef': None, 'namespace': '', 'function_pointer': 0, 'mutable': False, 'type': 'double', 'property_of_class': 'elem_t', 'parent': None, 'ctypes_type': 'ctypes.c_double', 'typedefs': 0, 'extern': False, 'class': 0, 'unresolved': False, 'name': 'double_a', 'fundamental': True}, {'line_number': 17, 'constant': 0, 'reference': 0, 'raw_type': 'std::string', 'static': 0, 'array': 0, 'pointer': 0, 'aliases': ['std::string'], 'typedef': None, 'namespace': '', 'function_pointer': 0, 'mutable': False, 'type': 'std::string', 'property_of_class': 'elem_t', 'parent': None, 'ctypes_type': 'ctypes.c_void_p', 'typedefs': 0, 'extern': False, 'class': 0, 'unresolved': True, 'name': 'string_a', 'fundamental': 0}], 'private': []}, 'typedefs': {'protected': [], 'public': [], 'private': []}, 'structs': {'protected': [], 'public': [], 'private': []}, 'enums': {'protected': [], 'public': [], 'private': []}, 'final': False, 'nested_classes': [], 'methods': {'protected': [], 'public': [{...}], 'private': []}}, 'defined': False, 'namespace': '', 'operator': False, 'static': False, 'returns_fundamental': True, 'rtnType': 'long', 'extern': False, 'path': 'elem_t', 'returns_pointer': 0, 'parameters': [{'raw_type': 'std::string', 'line_number': 18, 'typedef': None, 'unresolved': True, 'constant': 1, 'name': 'msg', 'parent': None, 'pointer': 0, 'ctypes_type': 'ctypes.c_void_p', 'function_pointer': 0, 'method': {...}, 'static': 0, 'fundamental': 0, 'mutable': False, 'extern': False, 'typedefs': 0, 'array': 0, 'type': 'const std::string &', 'class': 0, 'reference': 1, 'aliases': ['std::string']}], 'class': None, 'returns_reference': False, 'const': False, 'name': 'foomethod', 'pure_virtual': False, 'debug': '\t long foomethod ( const std::string & msg ) ;', 'explicit': False, 'virtual': False, 'destructor': False, 'returns': 'long', 'template': False, 'constructor': False, 'override': False, 'inline': False, 'final': False, 'friend': False, 'returns_class': False}
 #cpph.classes['elem_t']['methods']['public'][0].keys()
@@ -25,12 +27,7 @@ COMMENT = """/**
  */
 """
 
-METHOD_COMMENT_TEMPLATE = """/** 
- * @brief UNDOCUMENTED
- *
-{%for param in record['parameters']%} * @param {{param['name']}}
-{%endfor%}{%if record['returns'] != 'void'%} * @return {{record['returns']}}
-{%endif%} */"""
+METHOD_COMMENT_TEMPLATE = """"""
 
 class CommentTarget(object):
   
@@ -99,6 +96,10 @@ class CommentTarget(object):
     return "{0}:{1}".format(self.record['name'], self.line_number)
   
 
+def template_parse(template):
+  return
+  
+
 class ClassEntry(CommentTarget):
   def __init__(self, classRec, access=['public', 'protected', 'private']):
     CommentTarget.__init__(self, classRec)
@@ -118,7 +119,7 @@ class ClassEntry(CommentTarget):
 
 class FunctionEntry(CommentTarget):
   
-  template = j2.Template(METHOD_COMMENT_TEMPLATE)
+  template = None
   
   def __init__(self, methodRec):
     CommentTarget.__init__(self, methodRec)
@@ -136,30 +137,11 @@ class MethodEntry(FunctionEntry):
   def __init__(self, methodRec):
     FunctionEntry.__init__(self, methodRec)
     self.methodRec = methodRec
-    
-  
-
 
 class ClassMethodFile(object):
   def __init__(self, filename, targets):
     self.filename = filename
     self.targets = targets
-    
-def FileToLines(path):
-  lines = []
-  
-  f = open(path, "r")
-  for l in f:
-    lines.append(l)
-  
-  return lines
-
-
-def LinesToFile(lines, path):
-  f = open(path, "w")
-  for l in lines:
-    f.write(l)
-    
 
 
 def main():
@@ -168,6 +150,10 @@ def main():
   parser = argparse.ArgumentParser()
   
   path = "test_{0}".format(sys.argv[1])
+  
+  jenv = j2.Environment(loader=j2.FileSystemLoader(os.path.join(os.path.dirname(__file__), 'templates')))
+  
+  FunctionEntry.template = jenv.get_template('method.template')
   
   shutil.copy(sys.argv[1], path)
