+++ doxyboiler.wpr
@@ -4,8 +4,16 @@
 # Wing IDE project file                                          #
 ##################################################################
 [project attributes]
-proj.file-list = [loc('doxyboiler.py')]
+proj.directory-list = [{'dirloc': loc('templates'),
+                        'excludes': (),
+                        'filter': u'*',
+                        'include_hidden': False,
+                        'recursive': True,
+                        'watch_for_changes': True}]
+proj.file-list = [loc('doxyboiler.py'),
+                  loc('doxyboiler_utils.py')]
 proj.file-type = 'shared'
 proj.launch-config = {loc('doxyboiler.py'): ('project',
         (u'methods.h',
          ''))}
+proj.main-file = loc('doxyboiler.py')
